better version

+++ Mini-Golf/Mini-Golf/Obstacle.cpp
@@ -6,5 +6,5 @@ Obstacle::Obstacle(float X, float Y, float Width, float Height)
 	y = Y;
 	width = Width;
 	height = Height;
-	rect = { (float)x, (float)y, (float)width, (float)height };
+	rect = { x, y, width, height };
 }
