+++ Mini-Golf/Mini-Golf/Obstacle.h
@@ -6,7 +6,7 @@ class Obstacle
 public:
 	Obstacle(float X, float Y, float Width, float Height);
 
-	int x, y, width, height;
+	float x, y, width, height;
 	Rectangle rect; //testtt
 	
 private:
